Decompose LoginPage
