+++ arches_lingo/src/arches_lingo/types.ts
@@ -1,5 +1,13 @@
+import type { Ref } from "vue";
+
 export interface User {
     first_name: string;
     last_name: string;
     username: string;
+}
+
+// Prop injection types
+export interface UserRefAndSetter {
+    user: Ref<User | null>;
+    setUser: (userToSet: User | null) => void;
 }
